fix: Fix multiple compiler warnings

+++ gdx-jnigen-runtime/src/main/native/jni_env_tls.c
@@ -15,7 +15,7 @@
 
 #define HANDLE_RESULT(res, msg) \
     if (res != JNI_OK) { \
-        fprintf(stderr, msg ": %d", res); \
+        fprintf(stderr, msg ": %d", (int)res); \
         fflush(stderr); \
         exit(1); \
     }
